+++ src/Props/Editable.fs
@@ -6,34 +6,48 @@ open System
 
 [<Erase>]
 type editable =
-    /// Allows dynamic adjustment of deletion action based on row data.
+    /// Delete tooltip handler
+    static member inline deleteTooltip<'T> (handler: 'T -> string) = Interop.mkEditableAttr "deleteTooltip" handler
+    
+    /// Edit tooltip handler
+    static member inline editTooltip<'T> (handler: 'T -> string) = Interop.mkEditableAttr "editTooltip" handler
+
+    /// Determines if the row can be deleted
+    static member inline isDeletable<'T> (handler: 'T -> bool) =
+        Interop.mkEditableAttr "isDeletable" handler
+
+    /// Allows dynamic adjustment of deletion action based on row data
     static member inline isDeleteHidden<'T> (handler: 'T -> bool) =
         Interop.mkEditableAttr "isDeleteHidden" handler
+    
+    /// Determines if the row can be edited
+    static member inline isEditable<'T> (handler: 'T -> bool) =
+        Interop.mkEditableAttr "isEditable" handler
 
-    /// Allows dynamic adjustment of edit action based on row data.
+    /// Allows dynamic adjustment of edit action based on row data
     static member inline isEditHidden<'T> (handler: 'T -> bool) =
         Interop.mkEditableAttr "isEditHidden" handler
 
-    /// Fired when a new row is added, given row data.
+    /// Fired when a new row is added, given row data
     static member inline onRowAdd<'T> (handler: 'T -> JS.Promise<unit>) = 
         Interop.mkEditableAttr "onRowAdd" handler
-    /// Fired when a new row is added, given row data.
+    /// Fired when a new row is added, given row data
     static member inline onRowAdd<'T> (handler: 'T -> Async<unit>) = 
         Interop.mkEditableAttr "onRowAdd" (handler >> Async.StartAsPromise)
 
-    /// Fired when a row add is cancelled.
+    /// Fired when a row add is cancelled
     static member inline onRowAddCancelled<'T> (handler: 'T -> JS.Promise<unit>) = 
         Interop.mkEditableAttr "onRowAddCancelled" handler
-    /// Fired when a row add is cancelled.
+    /// Fired when a row add is cancelled
     static member inline onRowAddCancelled<'T> (handler: 'T -> Async<unit>) = 
         Interop.mkEditableAttr "onRowAddCancelled" (handler >> Async.StartAsPromise)
     
-    /// Fired when a row update is cancelled.
-    static member inline onRowUpdateCancelled<'T> (handler: 'T -> JS.Promise<unit>) = 
-        Interop.mkEditableAttr "onRowUpdateCancelled" handler
-    /// Fired when a row update is cancelled.
-    static member inline onRowUpdateCancelled<'T> (handler: 'T -> Async<unit>) = 
-        Interop.mkEditableAttr "onRowUpdateCancelled" (handler >> Async.StartAsPromise)
+    /// Fired when a row is deleted, given the old data
+    static member inline onRowDelete<'T> (handler: 'T -> JS.Promise<unit>) = 
+        Interop.mkEditableAttr "onRowDelete" handler
+    /// Fired when a row is deleted, given the old data
+    static member inline onRowDelete<'T> (handler: 'T -> Async<unit>) = 
+        Interop.mkEditableAttr "onRowDelete" (handler >> Async.StartAsPromise)
 
     /// Fired when a row is updated, given newData and oldData respectively
     static member inline onRowUpdate<'T> (handler: 'T -> 'T -> JS.Promise<unit>) = 
@@ -43,9 +57,9 @@ type editable =
         Func<_,_,_>(fun newData oldData -> handler newData oldData |> Async.StartAsPromise)
         |> Interop.mkEditableAttr "onRowUpdate"
 
-    /// Fired when a row is deleted, given the old data
-    static member inline onRowDelete<'T> (handler: 'T -> JS.Promise<unit>) = 
-        Interop.mkEditableAttr "onRowDelete" handler
-    /// Fired when a row is deleted, given the old data
-    static member inline onRowDelete<'T> (handler: 'T -> Async<unit>) = 
-        Interop.mkEditableAttr "onRowDelete" (handler >> Async.StartAsPromise)
+    /// Fired when a row update is cancelled
+    static member inline onRowUpdateCancelled<'T> (handler: 'T -> JS.Promise<unit>) = 
+        Interop.mkEditableAttr "onRowUpdateCancelled" handler
+    /// Fired when a row update is cancelled
+    static member inline onRowUpdateCancelled<'T> (handler: 'T -> Async<unit>) = 
+        Interop.mkEditableAttr "onRowUpdateCancelled" (handler >> Async.StartAsPromise)